suggesting 3d view and a docker deployment.
moved the visual display to a more 3 dimensoinal space.
this resulted in increasing free phone screen real estate.
this also makes it so the form is simpler, it gets switched instead of globally having
3 differnt forms on the DOM at the same time. this made it tricky to
work with the radio buttons tbh, now it should be easier.

using docker compose you will be able to connect your db and server closer.
and easier to do it too.